% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePhoto.R
\name{annotatePhoto}
\alias{annotatePhoto}
\title{Helper function to annotate a photo with the set of texts}
\usage{
annotatePhoto(photo_path, artist_text, date_created, caption_abstract)
}
\arguments{
\item{photo_path}{path to the foto}

\item{artist_text}{photo author name}

\item{date_created}{year when the photo was made}

\item{caption_abstract}{foto description}
}
\value{
returns foto with text
}
\description{
Helper function to annotate a photo with the set of texts
}
