% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nc.R
\name{check_nc}
\alias{check_nc}
\title{Functions for checking if netcdf file is available}
\usage{
check_nc(file_name = "israel_caves-2024.nc")
}
\arguments{
\item{file_name}{nc file name you're gonna use}
}
\value{
Boolean True or False
}
\description{
Functions for checking if netcdf file is available
}
\examples{
filename <- "israel_caves-2024.nc"
check_nc(filename)
}
