% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstcheck.R
\name{check_netcdf_file}
\alias{check_netcdf_file}
\title{Functions for checking if netcdf file is available}
\usage{
check_netcdf_file(file_name = "israel_caves-2024.nc")
}
\arguments{
\item{file_name}{}
}
\value{
No return, just info
}
\description{
Functions for checking if netcdf file is available
}
\examples{
check_netcdf_file("israel_caves-2024.nc")
}
