% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\docType{data}
\name{.ICCP_env}
\alias{.ICCP_env}
\title{Launches a Shiny app to explore Israeli cave climate data.}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.ICCP_env
}
\value{
This function is primarily invoked for its side effect of launching a Shiny app.
It does not return a value but will run the app until it is closed by the user.
}
\description{
Launches a Shiny app to explore Israeli cave climate data.
}
\examples{
\dontrun{
launchApp()
}

}
\keyword{datasets}
