% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedShiny.R
\name{feedShiny}
\alias{feedShiny}
\title{Function retreiving the data from a netcdf file}
\usage{
feedShiny(file_name = "israel_caves-2025.nc")
}
\arguments{
\item{file_name}{netcdf file path/name}
}
\value{
returns a list of data frames
}
\description{
Function retreiving the data from a netcdf file
}
\examples{
filename <- "israel_caves-2025.nc"
feedShiny(filename)

}
