% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGam.R
\name{fitGam}
\alias{fitGam}
\title{Removes seasonal component via GAM fitting}
\usage{
fitGam(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame
}
\description{
Removes seasonal component via GAM fitting
}
\examples{
set.seed(123)
example_data <- data.frame(
 datetime = seq.POSIXt(from = as.POSIXct("2024-01-01"), by = "day", length.out = 365),
 val = sin(seq(1, 365) * 2 * pi / 365) + rnorm(365, sd = 0.5),
 cave_name = sample(c("CaveA", "CaveB"), 365, replace = TRUE),
 var = sample(c("Temp", "Humidity"), 365, replace = TRUE),
 zone = sample(c("Zone1", "Zone2"), 365, replace = TRUE)
)
fitGam(example_data)

}
