% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSystime.R
\name{formatSystime}
\alias{formatSystime}
\title{Helper function for returning system time in a specific time format}
\usage{
formatSystime()
}
\value{
time value in a specific time format
}
\description{
Helper function for returning system time in a specific time format
}
\examples{
formatSystime()
}
