% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeta.R
\name{getMeta}
\alias{getMeta}
\title{Helper function to fetch specific metadata from foto files}
\usage{
getMeta(path)
}
\arguments{
\item{path}{Path to foto file}
}
\value{
data frame
}
\description{
Helper function to fetch specific metadata from foto files
}
