% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launches a Shiny app to explore Israeli cave climate data.}
\usage{
launchApp()
}
\value{
This function is primarily invoked for its side effect of launching a Shiny app.
It does not return a value but will run the app until it is closed by the user.
}
\description{
Launches a Shiny app to explore Israeli cave climate data.
}
\examples{
\dontrun{
launchApp()
}

}
