% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showFileDescription.r
\name{showFileDescription}
\alias{showFileDescription}
\title{Helper function to return Shiny modal window with a file details of a specific file}
\usage{
showFileDescription(file_path)
}
\arguments{
\item{file_path}{the path of the file}
}
\description{
Helper function to return Shiny modal window with a file details of a specific file
}
