% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstcheck.R
\name{check_netcdf_file}
\alias{check_netcdf_file}
\title{Functions for checking if netcdf file is available}
\usage{
check_netcdf_file(file_name = "israel_caves-2024.nc")
}
\arguments{
\item{file_name}{nc file name you're gonna use}
}
\value{
Boolean True or False
}
\description{
Functions for checking if netcdf file is available
}
\examples{
filename <- "israel_caves-2024.nc"
check_netcdf_file(filename)
}
